refactor: 네트워크 그래프 관련 타입 파일 이동 및 Network 타입 추가
