+++ types/index.ts
@@ -1,2 +1,3 @@
 export * from './artists';
+export * from './networkGraph';
 export * from './token';
